core: Move locks, atomic rc to unstable::sync
